requirements.txt
gymnasium==0.29.1
stable-baselines3==2.3.2
numpy==1.26.4
pandas==2.2.2
matplotlib==3.9.0
streamlit==1.37.1
plotly==5.24.1
PyYAML==6.0.2
pytest==8.3.3
tensorboard==2.18.0
